More advanced json serialization example.
